requirements.txt
line-bot-sdk==1.13.0
gunicorn==19.9.0
flask==1.1.1